refactor: Data and type separation
