+++ app/(club)/club/review/acceptance/page.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from "react";
 import ClubInfoSection from "../../content/clubInfoSection";
 import AcceptanceReviewDropdown from "@/components/dropdown/acceptanceReviewDropdown";
-import { ACCEPTANCE_REVIEWS } from "@/data/club";
+import { ACCEPTANCE_REVIEWS, MONILE_MENU_OPTIONS } from "@/data/club";
 import PlusBtn from "@/components/button/withIconBtn/plusBtn";
 import PointStatusBar from "@/components/bar/pointStatusBar";
 import MobilePointStatusBar from "@/components/bar/mobilePointStatusBar";
@@ -19,16 +19,6 @@ import AcceptanceReviewBottomSheet from "@/components/bottomSheet/acceptanceRevi
 import Alert from "@/components/alert/alert";
 import useResponsive from "../../../../../hooks/useResponsive";
 import AcceptanceReviewModal from "@/components/modal/review/acceptanceReviewModal";
-
-const options = [
-  { id: 0, label: "모집관리" },
-  { id: 1, label: "동아리 멤버 관리" },
-  { id: 2, label: "활동관리" },
-  { id: 3, label: "활동내역" },
-  { id: 4, label: "FAQ/Q&A" },
-  { id: 5, label: "활동후기" },
-  { id: 6, label: "합격후기" },
-];
 
 const ReviewPage = () => {
   const router = useRouter();
@@ -66,7 +56,7 @@ const ReviewPage = () => {
             </div>
             <div className="flex overflow-x-auto no-scrollbar ml-2 relative">
               <div className="flex whitespace-nowrap">
-                <SubTap optionData={options} />
+                <SubTap optionData={MONILE_MENU_OPTIONS} />
               </div>
             </div>
             <div className="flex items-center">
