+++ app/components/modal/review/acceptanceReviewModal.tsx
@@ -2,6 +2,7 @@ import Image from "next/image";
 import React, { useEffect, useState } from "react";
 import close from "@/images/icon/close.svg";
 import Alert from "@/components/alert/alert";
+import { INTERVIEWE_STYLE, INTERVIEWER } from "@/data/review";
 import ProgressBar from "@/components/bar/progressBar";
 import SmallBtn from "@/components/button/basicBtn/smallBtn";
 import DeleteBtn from "@/components/button/iconBtn/deleteBtn";
@@ -9,36 +10,13 @@ import RadioBtn from "@/components/button/radioBtn";
 import CustomInput from "@/components/input/customInput";
 import PullDown from "@/components/pulldown/pullDown";
 import CustomTextArea from "@/components/textArea/customTextArea";
-
-export const INTERVIEWER = [
-  { id: 0, label: "면접인원" },
-  { id: 1, label: "0명" },
-  { id: 2, label: "1명" },
-  { id: 3, label: "2명" },
-  { id: 4, label: "3명" },
-  { id: 5, label: "4명" },
-  { id: 6, label: "5명" },
-  { id: 7, label: "5명 이상" },
-];
-
-export const INTERVIEWE_STYLE = [
-  { id: 0, label: "면접방식" },
-  { id: 1, label: "온라인" },
-  { id: 2, label: "오프라인" },
-  { id: 3, label: "기타" },
-];
-
-interface AcceptanceReviewModalProps {
-  onClose: () => void;
-  onSubmit: () => void;
-  className?: string;
-}
+import { AcceptanceReviewProps } from "@/types/components/review";
 
 const AcceptanceReviewModal = ({
   onClose,
   onSubmit,
   className,
-}: AcceptanceReviewModalProps) => {
+}: AcceptanceReviewProps) => {
   const [title, setTitle] = useState<string>("");
   const [isVisible, setIsVisible] = useState<boolean>(false);
   const [interviewer, setInterviewer] = useState<string[]>([]);
