+++ app/data/review.ts
@@ -1,0 +1,17 @@
+export const INTERVIEWER = [
+  { id: 0, label: "면접인원" },
+  { id: 1, label: "0명" },
+  { id: 2, label: "1명" },
+  { id: 3, label: "2명" },
+  { id: 4, label: "3명" },
+  { id: 5, label: "4명" },
+  { id: 6, label: "5명" },
+  { id: 7, label: "5명 이상" },
+];
+
+export const INTERVIEWE_STYLE = [
+  { id: 0, label: "면접방식" },
+  { id: 1, label: "온라인" },
+  { id: 2, label: "오프라인" },
+  { id: 3, label: "기타" },
+];
