+++ app/types/components/review.ts
@@ -1,0 +1,14 @@
+export interface ActivityReviewProps {
+  onClose: () => void;
+  onSubmit: (data: {
+    title: string;
+    details: string;
+    badges: number[];
+  }) => void;
+}
+
+export interface AcceptanceReviewProps {
+  onClose: () => void;
+  onSubmit: () => void;
+  className?: string;
+}
